fix: high cyclomatic complexity
